move cli.rs, config.rs to cli/
