+++ src/cli/mod.rs
@@ -1,0 +1,9 @@
+#[allow(clippy::module_inception)]
+mod cli;
+pub use cli::*;
+
+mod config;
+pub use config::*;
+
+mod subcommand;
+pub use subcommand::*;
