+++ src/commands/alias.rs
@@ -1,4 +1,4 @@
-use crate::config::Config;
+use crate::cli::Config;
 use snm_core::{
     linker::Linker, version::dist_version::DistVersion, version::user_version::UserVersion, SnmRes,
 };
