+++ src/commands/exec.rs
@@ -1,4 +1,4 @@
-use crate::config::Config;
+use crate::cli::Config;
 use clap::Clap;
 use snm_core::{
     version::{dist_version::DistVersion, user_version::UserVersion},
