+++ src/commands/ls_remote.rs
@@ -1,4 +1,4 @@
-use crate::config::Config;
+use crate::cli::Config;
 use clap::Clap;
 use console::style;
 use snm_core::{fetcher::Fetcher, version::user_version::UserVersion, SnmRes};
