+++ src/commands/lts.rs
@@ -1,4 +1,4 @@
-use crate::config::Config;
+use crate::cli::Config;
 use clap::Clap;
 use console::style;
 use snm_core::{downloader::Downloader, fetcher::Fetcher, linker::Linker, SnmRes};
